Clean AppButton file, Update button and save icons to MCI and AntDesign (scalable, less blurry, and uses less space), refactor the UploadImage stuff into a separate file, allow non-editable image
